src/main/webapp/app/entities/departamentos/route/departamentos-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IDepartamentos, Departamentos } from '../departamentos.model';
import { DepartamentosService } from '../service/departamentos.service';

import { DepartamentosRoutingResolveService } from './departamentos-routing-resolve.service';

describe('Departamentos routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: DepartamentosRoutingResolveService;
  let service: DepartamentosService;
  let resultDepartamentos: IDepartamentos | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(DepartamentosRoutingResolveService);
    service = TestBed.inject(DepartamentosService);
    resultDepartamentos = undefined;
  });

  describe('resolve', () => {
    it('should return IDepartamentos returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDepartamentos = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDepartamentos).toEqual({ id: 123 });
    });

    it('should return new IDepartamentos if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDepartamentos = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultDepartamentos).toEqual(new Departamentos());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Departamentos })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDepartamentos = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDepartamentos).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
